Additional refactoring and model creation

+++ app-core/src/main/java/co/anitrend/support/crunchyroll/core/naviagation/NavigationTargets.kt
@@ -96,7 +96,7 @@ object NavigationTargets {
 
         @Parcelize
         data class Payload(
-            val mediaId: Int,
+            val mediaId: Long,
             val episodeThumbnail: String?
         ) : Parcelable
     }
