+++ app-domain/src/main/kotlin/co/anitrend/support/crunchyroll/domain/stream/models/MediaStreamQuery.kt
@@ -17,5 +17,5 @@
 package co.anitrend.support.crunchyroll.domain.stream.models
 
 data class MediaStreamQuery(
-    val mediaId: Int
+    val mediaId: Long
 )
